Address review comment to put tests in new class
